DESC: <feat><Project> 7 Segment Project added
BODY: This sample demonstrates the use of GPIO and 7-segment display control. A 2-digit 7-segment display is connected to the MCU. Pins A8, A9, and A10 are connected to LEDs that blink during operation. When the user presses the button connected to A1, counting starts from 0 to 99 on the 7-segment display. Pressing the button connected to A2 stops the counting.